application/requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
google-cloud-language==1.3.0
google-cloud-storage==1.31.2
grpcio==1.32.0
html5lib==1.0.1
ipaddr==2.2.0
lockfile==0.12.2
matplotlib-terminal==0.1a2
msgpack==0.6.2
pep517==0.8.2
progress==1.5
pylint==2.6.0
PySocks==1.7.1
pytoml==0.1.21
PyYAML==5.3.1
retrying==1.3.3
tweepy==3.9.0
webencodings==0.5.1